refactor: move checkIfLoggedIn to reusable function for other protected paths
